requirements.txt
flask==2.3.2
python-dotenv==1.0.0
pymongo[srv]
Flask-RESTful==0.3.10
pydantic==1.10.9
pandas==2.0.1
scikit-learn==1.3.0
redis==4.6.0
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0

fastapi[all]
uvicorn[standard]
python-jose[cryptography]